refactor(test): rename classes to implement FileProvider
