+++ app/src/test/java/fr/carbonit/th/provider/FileProvider.java
@@ -1,0 +1,7 @@
+package fr.carbonit.th.provider;
+
+import java.io.File;
+
+public interface FileProvider <T> {
+    File provide(T fileType);
+}
